Refactor: added validation to auth DTO
