+++ src/main/java/com/mudda/backend/auth/AuthRequest.java
@@ -8,9 +8,15 @@
  */
 package com.mudda.backend.auth;
 
-//TODO: add validation
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+
 public record AuthRequest(
+        @NotBlank
         String username,
+
+        @NotBlank
+        @Size(min = 8, max = 64)
         String password
 ) {
 }
