+++ src/main/java/com/mudda/backend/auth/RefreshRequest.java
@@ -8,7 +8,9 @@
  */
 package com.mudda.backend.auth;
 
+import jakarta.validation.constraints.NotBlank;
+
 public record RefreshRequest(
-        String refreshToken
+        @NotBlank String refreshToken
 ) {
 }
