+++ src/main/java/com/mudda/backend/auth/VerifyRequest.java
@@ -6,9 +6,12 @@
  * Created : 14-01-2026
  * ---------------------------------------------------------------
  */
-package com.mudda.backend.account;
+package com.mudda.backend.auth;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
 
 public record VerifyRequest(
-        String email
+        @NotBlank @Email String email
 ) {
 }
